requirements.txt
faiss-cpu
sentence-transformers
numpy
google-generativeai
